requirements.txt
Django==3.2.13
pytz==2020.1
psycopg2==2.9.5
django-bootstrap5==21.3
django-crispy-forms==1.14.0
django-appconf==1.0.5
pandas==1.3.5
Pillow==9.4.0
reportlab==3.6.12
django-dbbackup==4.0.2
django-mptt==0.14.0